+++ src/vybe/c.clj
@@ -899,7 +899,7 @@ signal(SIGSEGV, sighandler);
 
          {:keys [c-code form-hash schemas final-form]}
          (-> code-form
-             (transpile (assoc opts ::version 5)))
+             (transpile (assoc opts ::version 9)))
 
          obj-name (str "vybe_" sym-name "_"
                        (when (or (:no-cache sym-meta)
@@ -957,12 +957,12 @@ signal(SIGSEGV, sighandler);
                                                       (str/join ":"))))))}
                    ;; For safeness
                    ;; https://clang.llvm.org/docs/UndefinedBehaviorSanitizer.html#ubsan-checks
-                   (let [safe-flags ["-fsanitize=undefined"
-                                     "-fno-omit-frame-pointer"
+                   (let [safe-flags [#_"-fsanitize=undefined"
+                                     #_"-fno-omit-frame-pointer"
                                      "-g"]]
-                     (proc/sh {:env {"UBSAN_OPTIONS" "print_stacktrace=1"
-                                     "ASAN_SAVE_DUMPS" "MyFileName.dmp"
-                                     "PATH" (System/getenv "PATH")}}
+                     (proc/sh {:env {#_ #_"UBSAN_OPTIONS" "print_stacktrace=1"
+                                     #_ #_"ASAN_SAVE_DUMPS" "MyFileName.dmp"
+                                     #_ #_"PATH" (System/getenv "PATH")}}
                               (format
                                (->> (concat
                                      ["clang"